SCRUM 210 refactor: refactoring Payments component with table and pagination functionality
