+++ src/features/profile/settings/index.ts
@@ -1,4 +1,3 @@
 export { AccountManagement } from './ui/AccountManagement'
 export { Devices } from './ui/Devices'
 export { GeneralInfo } from './ui/General/GeneralInfo'
-export { Payments } from './ui/Payments/Payments'
