--- src/features/profile/settings/ui/Payments/Payments.module.scss
@@ -1,21 +1,0 @@
-.wrapper {
-  display: grid;
-  grid-template-rows: minmax(0, 1fr) auto;
-  gap: 1rem;
-  padding: 1.5rem 0;
-  height: calc(100vh - 220px);
-  min-height: 100%;
-  overflow: hidden;
-  position: relative;
-}
-
-.tableArea {
-  overflow: auto;
-  min-height: 0;
-}
-
-.table thead {
-  position: sticky;
-  top: 0px;
-  z-index: 2;
-}
