+++ src/features/subscriptions/model/index.ts
@@ -1,2 +1,4 @@
 export * from './mapSubscriptionTypeToLabel'
 export * from './mapPaymentTypeToLabel'
+export * from './payments.constants'
+export * from './payments.types'
