+++ src/features/subscriptions/model/payments.types.ts
@@ -1,0 +1,6 @@
+import { PaymentsSortBy, PaymentsSortDirection } from '@/shared/types'
+
+export type PaymentsSortState = {
+  key: PaymentsSortBy | null
+  direction: PaymentsSortDirection | null
+}
