+++ src/features/subscriptions/ui/Payments/PaymentsTable/PaymentsTable.module.scss
@@ -1,0 +1,10 @@
+.wrapper {
+  overflow: auto;
+  min-height: 0;
+
+  thead {
+    position: sticky;
+    z-index: 2;
+    top: 0px;
+  }
+}
