+++ src/features/subscriptions/ui/Payments/PaymentsTable/index.ts
@@ -1,0 +1,1 @@
+export * from './PaymentsTable'
