+++ src/features/subscriptions/ui/Payments/index.ts
@@ -1,0 +1,1 @@
+export * from './Payments'
