+++ src/features/subscriptions/ui/index.ts
@@ -1,1 +1,2 @@
 export * from './SubscriptionPricing'
+export * from './Payments'
