+++ src/shared/ui/Pagination/index.ts
@@ -1,0 +1,2 @@
+export { Pagination } from '@ictroot/ui-kit'
+export type { PaginationProps, PageItem } from '@ictroot/ui-kit'
