+++ src/shared/ui/index.ts
@@ -16,3 +16,4 @@ export * from './Modal'
 export * from './Separator'
 export * from './DatePicker'
 export * from './Select'
+export * from './Pagination'
